LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-expected.txt

CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'SomeOtherName' because it violates the following Content Security Policy directive: "trusted-types SomeName JustOneMoreName AnotherName"
CONSOLE MESSAGE: Refused to connect to http://localhost:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'AnotherName' because it violates the following Content Security Policy directive: "trusted-types SomeName JustOneMoreName AnotherName"
CONSOLE MESSAGE: Refused to connect to http://localhost:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS Allowed-name policy creation works.
PASS Another allowed-name policy creation works.
PASS Non-allowed name policy creation throws.
PASS Duplicate name policy creation throws.
